+++ Sources/cmdshelf/Error.swift
@@ -1,11 +1,11 @@
 import Foundation
 
 struct CmdshelfError: Error, CustomStringConvertible {
-    let message: String
-    init(_ message: String) {
+    let message: String?
+    init(_ message: String? = nil) {
         self.message = message
     }
     var description: String {
-        return message
+        return message ?? ""
     }
 }